+++ clickstream/clickstream-app/src/main/java/com/griddynamics/productindexer/model/ProductSearchResponse.java
@@ -16,5 +16,5 @@ public class ProductSearchResponse {
     private String currencyCode;
     private String category;
     private Map<String, String> attributes;
-    private Integer popularity;
+    private Double popularity;
 }
